feat: add bin roundtrip test + reorg tests
