+++ crates/league-toolkit/tests/bins/leona_small.bin
