--- crates/league-toolkit/tests/mod.rs
@@ -1,1 +1,0 @@
-pub mod meta;
